+++ packages/dbos-cloud/applications/get-app-logs.ts
@@ -1,6 +1,6 @@
 import axios, { AxiosError } from "axios";
 import { handleAPIErrors, getLogger, isCloudAPIErrorResponse, retrieveApplicationName } from "../cloudutils.js";
-import { loginGetCloudCredentials } from "../users/login.js";
+import { getCloudCredentials } from "../cloudutils.js";
 
 type LogResponse = {
   end: boolean;
@@ -24,7 +24,7 @@ export async function getAppLogs(host: string, last: number, pagesize: number, a
   }
 
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
   appName = appName || retrieveApplicationName(logger);
   if (!appName) {
