+++ packages/dbos-cloud/applications/list-apps.ts
@@ -1,11 +1,11 @@
 import axios, { AxiosError } from "axios";
 import { handleAPIErrors, getLogger, isCloudAPIErrorResponse } from "../cloudutils.js";
 import { Application, prettyPrintApplication } from "./types.js";
-import { loginGetCloudCredentials } from "../users/login.js";
+import { getCloudCredentials } from "../cloudutils.js";
 
 export async function listApps(host: string, json: boolean): Promise<number> {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
 
   try {
