+++ packages/dbos-cloud/dashboards/dashboards.ts
@@ -1,11 +1,11 @@
 import axios, { AxiosError } from "axios";
 import { getLogger, isCloudAPIErrorResponse, handleAPIErrors } from "../cloudutils.js";
 import open from "open";
-import { loginGetCloudCredentials } from "../users/login.js";
+import { getCloudCredentials } from "../cloudutils.js";
 
 export async function launchDashboard(host: string): Promise<number> {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
   try {
     logger.warn(`'dashboard launch' is a deprecated command; use 'dashboard url' instead.`);
@@ -42,7 +42,7 @@ export async function launchDashboard(host: string): Promise<number> {
 
 export async function getDashboardURL(host: string): Promise<number> {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
   try {
     const res = await axios.get(`https://${host}/v1alpha1/${userCredentials.organization}/dashboard`, {
@@ -75,7 +75,7 @@ export async function getDashboardURL(host: string): Promise<number> {
 
 export async function deleteDashboard(host: string): Promise<number> {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
   try {
     await axios.delete(`https://${host}/v1alpha1/${userCredentials.organization}/dashboard`, {
