+++ packages/dbos-cloud/organizations/organization.ts
@@ -1,10 +1,10 @@
 import axios, { AxiosError } from "axios";
 import { isCloudAPIErrorResponse, handleAPIErrors, getLogger } from "../cloudutils.js";
-import { loginGetCloudCredentials } from "../users/login.js";
+import { getCloudCredentials } from "../cloudutils.js";
 
 export async function orgInvite(host: string, json: boolean) {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
   try {
     const res = await axios.get(`https://${host}/v1alpha1/${userCredentials.organization}/secret`, {
@@ -39,7 +39,7 @@ export interface OrgUsers {
 
 export async function orgListUsers(host: string, json: boolean) {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
   try {
     const res = await axios.get(`https://${host}/v1alpha1/${userCredentials.organization}/users`, {
@@ -73,7 +73,7 @@ export async function orgListUsers(host: string, json: boolean) {
 
 export async function renameOrganization(host: string, oldname: string, newname: string) {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
 
   const currentOrg = userCredentials.organization;
@@ -113,7 +113,7 @@ export async function renameOrganization(host: string, oldname: string, newname:
 
 export async function joinOrganization(host: string, orgname: string, secret: string) {
   const logger = getLogger();
-  const userCredentials = await loginGetCloudCredentials(host, logger);
+  const userCredentials = await getCloudCredentials(host, logger);
   const bearerToken = "Bearer " + userCredentials.token;
 
   try {
