+++ packages/dbos-cloud/users/register.ts
@@ -1,109 +1,7 @@
-import axios, { AxiosError } from "axios";
-import { handleAPIErrors, getCloudCredentials, getLogger, isCloudAPIErrorResponse, credentialsExist, DBOSCloudCredentials, writeCredentials, deleteCredentials, UserProfile } from "../cloudutils.js";
-import validator from "validator";
-import { authenticate } from "./authentication.js";
-import { input } from "@inquirer/prompts";
+import { getCloudCredentials, getLogger } from "../cloudutils.js";
 
-function isValidUsername(username: string): boolean {
-  if (username.length < 3 || username.length > 30) {
-    return false;
-  }
-  return validator.matches(username, "^[a-z0-9_]+$");
-}
-
-export async function registerUser(username: string, secret: string, host: string): Promise<number> {
+export async function registerUser(username: string, secret: string | undefined, host: string): Promise<number> {
   const logger = getLogger();
-  let givenName = "";
-  let familyName = "";
-  let company = "";
-  if (!isValidUsername(username)) {
-    logger.error("Invalid username. Usernames must be between 3 and 30 characters long and contain only lowercase letters, underscores, and numbers.");
-    return 1;
-  }
-  if (!credentialsExist()) {
-    logger.info("Welcome to DBOS Cloud!");
-    logger.info("Before creating an account, please tell us a bit about yourself!");
-    givenName = await input({
-      message: "Enter first/given name:",
-      required: true,
-    });
-    familyName = await input({
-      message: "Enter last/family name:",
-      required: true,
-    });
-    company = await input({
-      message: "Enter company name:",
-      required: true,
-    });
-    const authResponse = await authenticate(logger);
-    if (authResponse === null) {
-      return 1;
-    }
-    const credentials: DBOSCloudCredentials = {
-      token: authResponse.token,
-      userName: username,
-      organization: "", // After auth0 auth only we don't have this information
-    };
-    writeCredentials(credentials);
-  } else {
-    const userCredentials = await getCloudCredentials();
-    if (userCredentials.userName !== username) {
-      logger.error(`You are trying to register ${username}, but are currently authenticated as ${userCredentials.userName}. Please run "npx dbos-cloud logout".`);
-      return 1;
-    } else {
-      logger.info(`You are currently authenticated as ${userCredentials.userName}. Registering ${userCredentials.userName} with DBOS Cloud...`);
-    }
-  }
-
-  const userCredentials = await getCloudCredentials();
-  const bearerToken = "Bearer " + userCredentials.token;
-  const loginName = userCredentials.userName;
-  try {
-    // Register the user in DBOS Cloud
-    await axios.put(
-      `https://${host}/v1alpha1/user`,
-      {
-        name: loginName,
-        given_name: givenName,
-        family_name: familyName,
-        company: company,
-        secret: secret,
-      },
-      {
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: bearerToken,
-        },
-      }
-    );
-    // Retrieve user profile and update local credentials
-    const response = await axios.get(`https://${host}/v1alpha1/user/profile`, {
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: bearerToken,
-      },
-    });
-    const profile = response.data as UserProfile;
-    const credentials: DBOSCloudCredentials = {
-      token: userCredentials.token,
-      refreshToken: userCredentials.refreshToken,
-      userName: profile.Name,
-      organization: profile.Organization,
-    };
-    writeCredentials(credentials);
-    logger.info(`${username} successfully registered!`);
-  } catch (e) {
-    const errorLabel = `Failed to register user ${loginName}`;
-    const axiosError = e as AxiosError;
-    if (isCloudAPIErrorResponse(axiosError.response?.data)) {
-      handleAPIErrors(errorLabel, axiosError);
-    } else {
-      logger.error(`${errorLabel}: ${(e as Error).message}`);
-    }
-    if (credentialsExist()) {
-      deleteCredentials();
-    }
-    return 1;
-  }
+  await getCloudCredentials(host, logger, username, secret);
   return 0;
 }
